Refactor to fix safari zooming issues
