+++ src/components/GalleryMedia.module.scss
@@ -4,6 +4,11 @@
   position: relative;
   width: 100%;
   height: 100%;
+
+  .transformWrapper, .transformContentClass {
+    width: 100% ;
+    height: 100%;
+  }
 }
 
 .media {
