+++ src/components/Slider.module.scss
@@ -30,10 +30,4 @@
 .fullscreen {
   height: 100%;
   width: 100%;
-
-  .transformWrapper, .transformContentClass {
-    width: 100% ;
-    height: 100%;
-  }
-
 }
